+++ src/simulation/constants.jl
@@ -67,8 +67,16 @@ const DEAD = 5
 const COMPARTMENTS = SA[
     UNINFECTED_NAIVE, UNINFECTED_IMMUNE,
     INFECTED_NAIVE, INFECTED_IMMUNE, DEAD
-    ]
+]
 const NUM_COMPARTMENTS = length(COMPARTMENTS)
 
-# Misc constants
+# Misc model constants
 const CHROMOSOME_SEPARATOR = "/"
+
+# Output constants
+const WITHIN_HOST_SEPARATOR = ";"
+const PARENT_SEPARATOR = "+"
+const COMPARTMENT_LABELS = [
+    "Uninfected Naive", "Uninfected Immune", "Infected Naive",
+    "Infected Immune", "Dead"
+]