+++ changelog.md
@@ -1,5 +1,20 @@
 # jOpqua Changelog
 
+## 19 February 2025
+- Finished composition dataframe (PCR)
+- Bug fix on history dataframe saving (PCR)
+
+## 18 February 2025
+- Added data csv export of compartments data (PCR)
+- Added csv plotting of compartments data (PCR)
+- Added option of `nothing` for `Response.matured.pathogen` (PCR)
+- Added data csv export of full data, can't believe it worked easily (PCR)
+- Change `Host` sampling for history to a random sample of hosts fixed over
+time (PCR)
+- Started work on composition dataframe creation (PCR)
+
+Next: composition dataframe and plot, then immunity
+
 ## 17 February 2025
 - Small bug fix: `mean_recombination_crossovers` type corrected (PCR)
 - Changed argument order in Type structs to put coefficients last within